+++ scw-mssql/Database/Logs/ApiLogQuery.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[ApiLogQuery] (
-   [ApiLogQueryID]					INTEGER IDENTITY(1,1)	NOT NULL,
-   [ApiLogID]						INTEGER					NOT NULL,
+   [ApiLogQueryID]					BIGINT IDENTITY(1,1)	NOT NULL,
+   [ApiLogID]						BIGINT					NOT NULL,
    [QueryType]						NVARCHAR(100)			NULL,
    [QueryValue]						NVARCHAR(400)			NULL,
    [LastUpdatedUTCDate]				DATETIME2				NOT NULL,
